fix: clamp agent logo icons in settings pickers to 14pt
SwiftUI Menu/Picker with .menu style renders items as NSMenuItems
and ignores SwiftUI frame sizing on custom icon views, drawing the
backing NSImage at its native pixel size. This caused the 256x256
logo PNGs to appear huge in pickers.

Apply the same NSImage-copy-with-clamped-size technique we use in
the tab bar sparkle menu (#244) to the two new pickers added in
PR #264:
- Settings > Agents > Default agent picker
- Settings > Changes > Agent picker

Extract the shared helper into AgentLogoView.menuImage(for:size:) so
TabBarView can also use it, deduplicating the icon clamping logic.
