+++ Alas/Sources/Settings/AgentsPane.swift
@@ -108,7 +108,7 @@ struct AgentsPane: View {
                 Label {
                     Text(agent.displayName)
                 } icon: {
-                    AgentLogoView(agent: agent, size: 14)
+                    Image(nsImage: AgentLogoView.menuImage(for: agent, size: 14))
                 }
                 .tag(agent.id)
             }
